# routers/api/docs/swagger.yaml
basePath: /
definitions:
  Rsp:
    properties:
      code:
        allOf:
        - $ref: '#/definitions/proto.CodeType'
        description: msg code
      message:
        description: msg info
        type: string
      requestId:
        description: trans id
        type: string
      results:
        description: response body
    type: object
  async.AsyncTask:
    properties:
      processed:
        type: integer
      taskId:
        type: string
      taskStatus:
        type: string
      total:
        type: integer
    type: object
  httputil.HTTPError:
    properties:
      code:
        example: 400
        type: integer
      message:
        example: status bad request
        type: string
    type: object
  proto.CodeType:
    enum:
    - 200
    - 201
    - 1001
    - 1002
    - 1003
    - 1004
    - 1005
    - 1006
    - 1007
    - 1008
    - 1009
    - 1010
    - 1011
    - 1012
    - 1013
    - 1014
    - 1015
    - 1016
    - 1017
    - 1018
    - 1021
    - 1022
    - 1027
    - 1028
    - 1029
    - 1030
    - 1031
    - 1032
    - 1036
    - 1037
    - 1038
    - 1046
    - 1050
    - 1061
    - 1062
    type: integer
    x-enum-comments:
      CodeCopyIdError: 文件夹不能拷贝到自己
      CodeFailedToConnect: 建立连接失败
      CodeFailedToConnectDB: 连接DB失败
      CodeFailedToCreateBucket: 失败去创建桶
      CodeFailedToCreateFolder: 失败去创建文件夹
      CodeFailedToCreateMultipartTask: 失败去创建任务
      CodeFailedToCreateSymlink: 失败去创建符号链接
      CodeFailedToDeleteUser: 删除用户失败
      CodeFailedToGetUsedStorage: 获取用户空间使用量失败
      CodeFailedToInitUser: 失败去初始化用户
      CodeFailedToOpenFile: 打开文件失败
      CodeFailedToOperateDB: 失败去操作数据库
      CodeFailedToOptMinio: 失败去写文件
      CodeFailedToSaveFile: 失败去保存数据
      CodeFileExist: 文件已存在
      CodeFileNotExist: 文件不存在
      CodeFolderDepthTooLong: 文件夹层数超过5层
      CodeFolderExist: 文件夹已存在
      CodeFolderNotExist: 文件夹不存在
      CodeGetAsyncTaskInfoFailed: 获取异步任务状态失败
      CodeMultipartRangeUploaded: 分片范围已上传
      CodeMultipartTaskCompleteErr: 合并错误
      CodeMultipartTaskHashErr: 文件hash校验错误
      CodeMultipartTaskNotFound: 未找到任务
      CodeMultipartTaskOverlap: 任务重叠
      CodeMultipartTaskRangeErr: 上传范围错误
      CodeMultipartUploadingConflit: 分片范围上传冲突
      CodeNotEnoughSpace: 空间不够，不上传
      CodeParamErr: 参数错误
      CodeReqParamErr: 请求参数错误(推荐)
      CodeUserIdError: 用户Id的错误
      CodeWriteRedisFailed: 写入redis 失败
      CodeZipFileFailed: 压缩文件失败
    x-enum-varnames:
    - CodeOk
    - CodeCreateAsyncTaskSuccess
    - CodeParamErr
    - CodeReqParamErr
    - CodeFileNotExist
    - CodeFailedToOpenFile
    - CodeFolderNotExist
    - CodeFailedToConnect
    - CodeFailedToConnectDB
    - CodeFailedToOptMinio
    - CodeFailedToCreateBucket
    - CodeFailedToSaveFile
    - CodeFailedToOperateDB
    - CodeFailedToCreateFolder
    - CodeFileExist
    - CodeFolderExist
    - CodeFolderDepthTooLong
    - CodeUserIdError
    - CodeFailedToInitUser
    - CodeFailedToDeleteUser
    - CodeFailedToGetUsedStorage
    - CodeCopyIdError
    - CodeFailedToCreateMultipartTask
    - CodeMultipartTaskNotFound
    - CodeMultipartTaskOverlap
    - CodeMultipartTaskRangeErr
    - CodeMultipartTaskHashErr
    - CodeMultipartTaskCompleteErr
    - CodeNotEnoughSpace
    - CodeMultipartRangeUploaded
    - CodeMultipartUploadingConflit
    - CodeWriteRedisFailed
    - CodeZipFileFailed
    - CodeFailedToCreateSymlink
    - CodeGetAsyncTaskInfoFailed
...
  proto.CopyFileReq:
    properties:
      dstPath:
        description: dest path
        type: string
      uuids:
        description: uuids
        items:
          type: string
        type: array
        uniqueItems: true
    type: object
  proto.CopyRsp:
    properties:
      affectRows:
        type: integer
      data:
        items:
          $ref: '#/definitions/proto.NewAndOldUuid'
        type: array
    type: object
...
  proto.DbAffect:
    properties:
      affectRows:
        type: integer
    type: object
  proto.DeleteFileReq:
    properties:
      uuids:
        description: uuids
        items:
          type: string
        type: array
        uniqueItems: true
    type: object
...
  proto.FileInfoForInnerRsp:
    properties:
      betag:
        type: string
      bucketName:
        type: string
      category:
        type: string
      createdAt:
        type: integer
      duration:
        type: integer
      executable:
        type: boolean
      ext:
        items:
          type: integer
        type: array
      fileCount:
        type: integer
      isDir:
        type: boolean
      mime:
        type: string
      modifyAt:
        type: integer
      name:
        type: string
      operationAt:
        type: integer
      parentUuid:
        type: string
      path:
        type: string
      relativePath:
        type: string
      size:
        type: integer
      tags:
        type: string
      transactionId:
        type: integer
      trashed:
        description: 0-normal; 1-Logical delete, put into the recycle bin; 2-Has been
          cleared from the recycle bin and is to be physically deleted
        type: integer
      userId:
        type: integer
      uuid:
        type: string
      version:
        type: integer
    type: object
...
  proto.FileInfoForTrendsReq:
    properties:
      uuids:
        description: uuids
        items:
          type: string
        type: array
    type: object
  proto.FileInfoForTrendsRsp:
    properties:
      fileInfos:
        items:
          $ref: '#/definitions/proto.FileInfoForTrends'
        type: array
    type: object
  proto.FileInfoPub:
    properties:
      betag:
        type: string
      category:
        type: string
      createdAt:
        type: integer
      fileCount:
        type: integer
      isDir:
        type: boolean
      mime:
        type: string
      modifyAt:
        type: integer
      name:
        type: string
      operationAt:
        type: integer
      parentUuid:
        type: string
      path:
        type: string
      size:
        type: integer
      trashed:
        description: 0-normal; 1-Logical delete, put into the recycle bin; 2-Has been
          cleared from the recycle bin and is to be physically deleted
        type: integer
      uuid:
        type: string
    type: object
  proto.FileInfoRsp:
    properties:
      betag:
        type: string
      category:
        type: string
      createdAt:
        type: integer
      fileCount:
        type: integer
      isDir:
        type: boolean
      mime:
        type: string
      modifyAt:
        type: integer
      name:
        type: string
      operationAt:
        type: integer
      parentUuid:
        type: string
      path:
        type: string
      size:
        type: integer
      trashed:
        description: 0-normal; 1-Logical delete, put into the recycle bin; 2-Has been
          cleared from the recycle bin and is to be physically deleted
        type: integer
      uuid:
        type: string
    type: object
...
  proto.GetListRspData:
    properties:
      fileList:
        items:
          $ref: '#/definitions/proto.FileInfoPub'
        type: array
      pageInfo:
        $ref: '#/definitions/proto.PageInfoExt'
    type: object
...
  proto.ModifyFileReq:
    properties:
      fileName:
        description: new name
        type: string
      uuid:
        description: file/folder uuid
        type: string
    type: object
  proto.MoveFileReq:
    properties:
      destPath:
        description: dest path
        type: string
      uuids:
        description: uuid
        items:
          type: string
        type: array
        uniqueItems: true
    required:
    - uuids
    type: object
  proto.NewAndOldUuid:
    properties:
      newId:
        type: string
      oldId:
        type: string
    type: object
  proto.PageInfoExt:
    properties:
      count:
        type: integer
      page:
        type: integer
      pageSize:
        type: integer
      total:
        type: integer
    type: object
...
  proto.VodSymlinkReq:
    properties:
      uuid:
        description: file's uuid
        type: string
    type: object
  proto.VodSymlinkRsp:
    properties:
      linkName:
        type: string
    type: object
info:
  contact:
    email: service@ao.space
    name: AO.space
    url: https://ao.space/
  description: This is AO.space aofs OpenAPI reference document.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: aofs apis
  version: "1.0"
paths:
...
  /space/v1/api/file/copy:
    post:
      consumes:
      - application/json
      description: Batch copy files
      parameters:
      - description: user id
        in: query
        name: userId
        required: true
        type: string
      - description: params
        in: body
        name: copyFilesReq
        required: true
        schema:
          $ref: '#/definitions/proto.CopyFileReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/Rsp'
            - properties:
                results:
                  $ref: '#/definitions/proto.CopyRsp'
              type: object
      summary: Batch copy files
      tags:
      - File
  /space/v1/api/file/delete:
    post:
      consumes:
      - application/json
      description: Delete files/folders to Recycle bin
      parameters:
      - description: user id
        in: query
        name: userId
        required: true
        type: string
      - description: params
        in: body
        name: DeleteId
        required: true
        schema:
          $ref: '#/definitions/proto.DeleteFileReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Rsp'
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/Rsp'
            - properties:
                results:
                  $ref: '#/definitions/async.AsyncTask'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httputil.HTTPError'
      summary: Delete files/folders to Recycle bin
      tags:
      - File
...
  /space/v1/api/file/info:
    get:
      consumes:
      - application/json
      description: Query file info
      parameters:
      - description: user id
        in: query
        name: userId
        required: true
        type: integer
      - description: file/folder uuid
        in: query
        name: uuid
        type: string
      - description: path
        in: query
        name: path
        type: string
      - description: name
        in: query
        name: name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: file info
          schema:
            allOf:
            - $ref: '#/definitions/Rsp'
            - properties:
                results:
                  $ref: '#/definitions/proto.FileInfoRsp'
              type: object
      summary: Query file info
      tags:
      - File
  /space/v1/api/file/list:
    get:
      consumes:
      - application/json
      description: Get file list
      parameters:
      - description: user id
        in: query
        name: userId
        required: true
        type: integer
      - description: folder uuid
        in: query
        name: uuid
        type: string
      - description: Whether to filter folders
        in: query
        name: isDir
        type: boolean
      - description: page，default:1
        in: query
        name: page
        type: integer
      - description: page size，default:10
        in: query
        name: pageSize
        type: integer
      - description: Sort. The default is reverse order
        in: query
        name: orderBy
        type: string
      - description: 'file classification, field value: document，video，picture or
          other; If there is no field, all are included'
        in: query
        name: category
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/Rsp'
            - properties:
                results:
                  $ref: '#/definitions/proto.GetListRspData'
              type: object
      summary: Get file list
      tags:
      - File
  /space/v1/api/file/move:
    post:
      consumes:
      - application/json
      description: Move file/folder
      parameters:
      - description: user id
        in: query
        name: userId
        required: true
        type: string
      - description: params
        in: body
        name: moveFilesReq
        required: true
        schema:
          $ref: '#/definitions/proto.MoveFileReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/Rsp'
            - properties:
                results:
                  $ref: '#/definitions/proto.DbAffect'
              type: object
      summary: Move file/folder
      tags:
      - File
  /space/v1/api/file/rename:
    post:
      consumes:
      - application/json
      description: Modify file/folder name
      parameters:
      - description: user id
        in: query
        name: userId
        required: true
        type: string
      - description: request info
        in: body
        name: ModifyFileReq
        required: true
        schema:
          $ref: '#/definitions/proto.ModifyFileReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/Rsp'
            - properties:
                results:
                  $ref: '#/definitions/proto.DbAffect'
              type: object
      summary: Modify file/folder name
      tags:
      - File
  /space/v1/api/file/search:
    get:
      consumes:
      - application/json
      description: Search files
      parameters:
      - description: user id
        in: query
        name: userId
        required: true
        type: string
      - description: 'folder''s uuid，default: /'
        in: query
        name: uuid
        type: string
      - description: filename
        in: query
        name: name
        required: true
        type: string
      - description: file type
        in: query
        name: category
        type: string
      - description: page, default:1
        in: query
        name: page
        type: integer
      - description: page size，default:10
        in: query
        name: pageSize
        type: integer
      - description: sort type, default value is in reverse order of change time
        in: query
        name: orderBy
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/Rsp'
            - properties:
                results:
                  $ref: '#/definitions/proto.GetListRspData'
              type: object
      summary: Search files
      tags:
      - File
...
  /space/v1/api/file/vod/symlink:
    post:
      consumes:
      - application/json
      description: Create symbolic link
      parameters:
      - description: user id
        in: query
        name: userId
        required: true
        type: string
      - description: parmas
        in: body
        name: VodSymlinkReq
        required: true
        schema:
          $ref: '#/definitions/proto.VodSymlinkReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/Rsp'
            - properties:
                results:
                  $ref: '#/definitions/proto.VodSymlinkRsp'
              type: object
      summary: Create symbolic link
      tags:
      - File
...
  /space/v1/api/inner/file/info:
    get:
      consumes:
      - application/json
      description: Get file info for inner
      parameters:
      - description: user id
        in: query
        name: userId
        required: true
        type: integer
      - description: uuid
        in: query
        name: uuid
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/Rsp'
            - properties:
                results:
                  $ref: '#/definitions/proto.FileInfoForInnerRsp'
              type: object
      summary: Get file info for inner
      tags:
      - File
  /space/v1/api/inner/file/infos:
    post:
      consumes:
      - application/json
      description: Get fileinfo list for inner
      parameters:
      - description: user id
        in: query
        name: userId
        required: true
        type: integer
      - description: uuids
        in: body
        name: FileInfoForTrendsReq
        schema:
          $ref: '#/definitions/proto.FileInfoForTrendsReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/Rsp'
            - properties:
                results:
                  $ref: '#/definitions/proto.FileInfoForTrendsRsp'
              type: object
      summary: Get fileinfo list for inner
      tags:
      - File
...
swagger: "2.0"